+++ src/api/consumer/index.js
@@ -16,6 +16,7 @@ import getComponentLogs from './lib/get-component-logs';
 import { add as remoteAdd, list as remoteList, remove as remoteRm } from './lib/remote';
 import config from './lib/global-config';
 import getDriver from './lib/get-driver';
+import watchInlineComponents from './lib/watch-inline-components';
 
 export {
   init,
@@ -40,4 +41,5 @@ export {
   remoteList,
   remoteRm,
   getDriver,
+  watchInlineComponents,
 };