+++ src/cli/command-registrar-builder.js
@@ -29,6 +29,7 @@ import RefreshScope from './commands/private-cmds/refresh-scope-cmd';
 import CatScope from './commands/private-cmds/cat-scope-cmd';
 import ScopeConfig from './commands/public-cmds/scope-config-cmd';
 import Bind from './commands/public-cmds/bind-cmd';
+import Watch from './commands/public-cmds/watch-cmd';
 
 export default function registerCommands(): CommandRegistrar {
   return new CommandRegistrar(BIT_USAGE, BIT_DESCRIPTION, BIT_VERSION, [
@@ -60,5 +61,6 @@ export default function registerCommands(): CommandRegistrar {
     new CatScope(),
     new ScopeConfig(),
     new Bind(),
+    new Watch(),
   ]);
 }