+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- enable bit watch command -> build-all-inline on every change
+
 ## [0.5.12] - 2017-05-14 rc
 
 - enable "bit build --inline" command with no arguments for building all inline components